// webapp/src/api/model/currencySettings.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CurrencyCalculationEntry } from './currencyCalculationEntry';

export interface CurrencySettings {
  baseCurrency: string;
  baseCurrencyShortForm: string;
  calculationEntries: CurrencyCalculationEntry[];
}